Combined tests for ChallengeInteraction and GridInteraction into tests for TileInteraction
